+++ plugins/deep-link/src/config.rs
@@ -7,7 +7,7 @@
 use serde::{Deserialize, Deserializer};
 use tauri_utils::config::DeepLinkProtocol;
 
-#[derive(Deserialize)]
+#[derive(Deserialize, Clone)]
 pub struct AssociatedDomain {
     #[serde(deserialize_with = "deserialize_associated_host")]
     pub host: String,
@@ -29,7 +29,7 @@ where
     }
 }
 
-#[derive(Deserialize)]
+#[derive(Deserialize, Clone)]
 pub struct Config {
     /// Mobile requires `https://<host>` urls.
     #[serde(default)]
@@ -41,7 +41,7 @@ pub struct Config {
     pub desktop: DesktopProtocol,
 }
 
-#[derive(Deserialize)]
+#[derive(Deserialize, Clone)]
 #[serde(untagged)]
 #[allow(unused)] // Used in tauri-bundler
 pub enum DesktopProtocol {
@@ -52,5 +52,28 @@ pub enum DesktopProtocol {
 impl Default for DesktopProtocol {
     fn default() -> Self {
         Self::List(Vec::new())
+    }
+}
+
+impl DesktopProtocol {
+    #[allow(dead_code)]
+    pub fn contains_scheme(&self, scheme: &String) -> bool {
+        match self {
+            Self::One(protocol) => protocol.schemes.contains(scheme),
+            Self::List(protocols) => protocols
+                .iter()
+                .any(|protocol| protocol.schemes.contains(scheme)),
+        }
+    }
+
+    #[allow(dead_code)]
+    pub fn schemes(&self) -> Vec<String> {
+        match self {
+            Self::One(protocol) => protocol.schemes.clone(),
+            Self::List(protocols) => protocols
+                .iter()
+                .flat_map(|protocol| protocol.schemes.clone())
+                .collect(),
+        }
     }
 }